CambiosKiwiPersona

+++ src/main/java/com/sisu/sisu/Dao/IPersonaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.Persona;
 
-public interface IPersonaDao extends CrudRepository<Persona, Long> {
+public interface IPersonaDao extends CrudRepository<Persona, Integer> {
     
 }
