+++ src/main/java/com/sisu/sisu/Service/IPersonaService.java
@@ -8,6 +8,6 @@ import com.sisu.sisu.entitys.Persona;
 public interface IPersonaService {
     public List<Persona> findAll(); //Muestra todos los registros
     public void save(Persona persona); //Guarda registro
-    public Persona findOne(Long id); //Muestra un registro
-    public void delete(Long id); //Elimina registro
+    public Persona findOne(Integer id); //Muestra un registro
+    public void delete(Integer id); //Elimina registro
 }
